refactor(stats): split drawdown and periodic-reporting into focused mixins (#840)
The two largest _stats modules conflated several concerns. Extract two
cohesive, self-contained mixins (each touching only self._data/self.all, no
cross-mixin calls):

- _DrawdownMixin (_drawdown.py) ← compsum, drawdown, prices,
  max_drawdown[_single_series], drawdown_details (from _performance.py)
- _PeriodicReportingMixin (_periodic.py) ← monthly_returns, distribution,
  compare, worst_n_periods (from _reporting.py)

Both are added to the Stats MRO, so the public API is unchanged. Internal
references to the moved staticmethod prices() (in _rolling.py and a migration
test) now point at _DrawdownMixin, and the _reporting cross-mixin stub for
max_drawdown is re-pointed.

_performance.py 900->713 lines, _reporting.py 830->581. All gates green;
coverage stays at 100%.
